third_party/blink/web_tests/http/tests/htmlimports/import-upgrade-iframe-expected.txt

CONSOLE WARNING: line 11: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
PASS target.isCreated is true
PASS successfullyParsed is true

TEST COMPLETE
